Use game config for easy tweak
