+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/four_of_a_kind.rs
@@ -15,7 +15,7 @@ fn test_four_of_a_kind() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::FourOfAKind);
     assert_eq!(template.rank, Rank::Ace);
 }
@@ -31,7 +31,7 @@ fn test_five_of_a_kind_is_treated_as_four_of_a_kind() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::FourOfAKind);
     assert_eq!(template.rank, Rank::Ace);
 }
