+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/high_card.rs
@@ -15,7 +15,7 @@ fn test_high_card() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::High);
     assert_eq!(template.rank, Rank::Ace);
 }
