+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/one_pair.rs
@@ -15,7 +15,7 @@ fn test_one_pair() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::OnePair);
     assert_eq!(template.rank, Rank::Ace);
 }
