+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/royal_flush.rs
@@ -15,7 +15,7 @@ fn test_royal_flush() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Ace);
@@ -31,7 +31,7 @@ fn test_royal_flush_4cards_without_upgrade() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_ne!(template.kind, TowerKind::RoyalFlush);
 }
 
@@ -48,7 +48,7 @@ fn test_royal_flush_4cards_with_upgrade() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Ace);
@@ -67,7 +67,7 @@ fn test_royal_flush_skip_rank() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::High);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Ace);
@@ -87,7 +87,7 @@ fn test_royal_flush_skip_rank_and_shorten_4cards() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Ace);
@@ -107,7 +107,7 @@ fn test_royal_flush_treat_suits_as_same() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert!(template.suit == Suit::Hearts || template.suit == Suit::Diamonds);
     assert_eq!(template.rank, Rank::Ace);
@@ -127,7 +127,7 @@ fn test_royal_flush_treat_suits_as_same_and_shorten_4cards() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert!(template.suit == Suit::Hearts || template.suit == Suit::Diamonds);
     assert_eq!(template.rank, Rank::Ace);
@@ -148,7 +148,7 @@ fn test_royal_flush_treat_suits_as_same_and_shorten_4cards_and_skip_rank_for_str
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::RoyalFlush);
     assert!(template.suit == Suit::Hearts || template.suit == Suit::Diamonds);
     assert_eq!(template.rank, Rank::Ace);
