+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/straight.rs
@@ -15,7 +15,7 @@ fn test_straight() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Jack);
 }
@@ -30,7 +30,7 @@ fn test_straight_4cards_without_upgrade() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_ne!(template.kind, TowerKind::Straight);
 }
 
@@ -47,7 +47,7 @@ fn test_straight_4cards_with_upgrade() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Ten);
 }
@@ -65,7 +65,7 @@ fn test_straight_skip_rank() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::High);
     assert_eq!(template.rank, Rank::Jack);
 }
@@ -84,7 +84,7 @@ fn test_straight_skip_rank_and_shorten_4cards() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Jack);
 }
@@ -103,7 +103,7 @@ fn test_straight_with_removed_two_allows_ace_low() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Six);
 }
@@ -124,7 +124,7 @@ fn test_straight_with_removed_two_and_shorten_4cards_allows_ace_low() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Six);
 }
@@ -143,7 +143,7 @@ fn test_straight_with_removed_two_and_three_allows_ace_low() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Seven);
 }
@@ -162,7 +162,7 @@ fn test_straight_with_removed_two_still_recognizes_included_two() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Six);
 }
@@ -182,7 +182,7 @@ fn test_straight_with_removed_two_and_skip_rank_allows_ace_four_five_six_seven()
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::Straight);
     assert_eq!(template.rank, Rank::Seven);
 }
