+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tests/straight_flush.rs
@@ -15,7 +15,7 @@ fn test_straight_flush() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::King);
@@ -31,7 +31,7 @@ fn test_straight_flush_4cards_without_upgrade() {
     ];
     let upgrade_state = UpgradeState::default();
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_ne!(template.kind, TowerKind::StraightFlush);
 }
 
@@ -48,7 +48,7 @@ fn test_straight_flush_4cards_with_upgrade() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::King);
@@ -69,7 +69,7 @@ fn test_straight_flush_with_removed_two_and_shorten_4cards_allows_ace_low() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Six);
@@ -88,7 +88,7 @@ fn test_straight_flush_skip_rank() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::High);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Ace);
@@ -108,7 +108,7 @@ fn test_straight_flush_treat_suits_as_same() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert!(template.suit == Suit::Hearts || template.suit == Suit::Diamonds);
     assert_eq!(template.rank, Rank::King);
@@ -128,7 +128,7 @@ fn test_straight_flush_treat_suits_as_same_and_shorten_4cards() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert!(template.suit == Suit::Hearts || template.suit == Suit::Diamonds);
     assert_eq!(template.rank, Rank::King);
@@ -148,7 +148,7 @@ fn test_straight_flush_with_removed_two_still_recognizes_included_two() {
         ..UpgradeState::default()
     };
     let rerolled_count = 0;
-    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count);
+    let template = get_highest_tower_template(&cards, &upgrade_state, rerolled_count, &crate::config::GameConfig::default_config());
     assert_eq!(template.kind, TowerKind::StraightFlush);
     assert_eq!(template.suit, Suit::Hearts);
     assert_eq!(template.rank, Rank::Six);
