+++ tower-defense/src/flow_ui/selecting_tower/tower_selecting_hand/get_highest_tower/tower_template_factory.rs
@@ -1,16 +1,13 @@
 use crate::card::{Rank, Suit};
+use crate::config::GameConfig;
 use crate::game_state::tower::TowerKind;
 use crate::game_state::tower::TowerTemplate;
 
-pub fn create_tower_template(kind: TowerKind, suit: Suit, rank: Rank) -> TowerTemplate {
-    TowerTemplate {
-        kind,
-        shoot_interval: kind.shoot_interval(),
-        default_attack_range_radius: kind.default_attack_range_radius(),
-        default_damage: kind.default_damage(),
-        suit,
-        rank,
-        skill_templates: kind.skill_templates(),
-        default_status_effects: vec![],
-    }
+pub fn create_tower_template(
+    kind: TowerKind,
+    suit: Suit,
+    rank: Rank,
+    config: &GameConfig,
+) -> TowerTemplate {
+    TowerTemplate::new_with_config(kind, suit, rank, config)
 }
