+++ tower-defense/src/game_state/effect/mod.rs
@@ -135,7 +135,7 @@ pub fn run_effect_with_rng<R: rand::Rng + ?Sized>(
 ) {
     match effect {
         Effect::Heal { amount } => {
-            game_state.hp = (game_state.hp + amount).min(crate::game_state::MAX_HP);
+            game_state.hp = (game_state.hp + amount).min(game_state.config.player.max_hp);
         }
         Effect::Shield { amount } => {
             game_state.shield += amount;
@@ -193,7 +193,10 @@ pub fn run_effect_with_rng<R: rand::Rng + ?Sized>(
             game_state.upgrade_state.upgrade(upgrade);
         }
         Effect::GrantItem { rarity: _ } => {
-            let item = crate::game_state::item::generation::generate_item_with_rng(rng);
+            let item = crate::game_state::item::generation::generate_item_with_rng(
+                rng,
+                &game_state.config,
+            );
             game_state.items.push(item);
         }
         Effect::IncreaseAllTowersDamage { multiplier } => {
@@ -290,11 +293,14 @@ pub fn run_effect_with_rng<R: rand::Rng + ?Sized>(
         } => {
             for _ in 0..*count {
                 if matches!(game_state.flow, GameFlow::PlacingTower) {
-                    game_state.hand.push(HandItem::Tower(TowerTemplate::new(
-                        *tower_kind,
-                        *suit,
-                        *rank,
-                    )));
+                    game_state
+                        .hand
+                        .push(HandItem::Tower(TowerTemplate::new_with_config(
+                            *tower_kind,
+                            *suit,
+                            *rank,
+                            &game_state.config,
+                        )));
                 } else {
                     game_state
                         .stage_modifiers
@@ -317,7 +323,7 @@ pub fn run_effect_with_rng<R: rand::Rng + ?Sized>(
             max_amount,
         } => {
             let heal_amount = rng.gen_range(*min_amount..=*max_amount);
-            game_state.hp = (game_state.hp + heal_amount).min(crate::game_state::MAX_HP);
+            game_state.hp = (game_state.hp + heal_amount).min(game_state.config.player.max_hp);
         }
         Effect::GainGold {
             min_amount,
@@ -509,6 +515,7 @@ pub mod tests_support {
                 ),
             black_smoke_sources: Default::default(),
             base_animation_state: crate::game_state::BaseAnimationState::new(Instant::now()),
+            config: std::sync::Arc::new(crate::config::GameConfig::default_config()),
 
             hand_panel_forced_open: true,
             shop_panel_forced_open: true,
