+++ tower-defense/src/game_state/monster/monster_template.rs
@@ -18,83 +18,32 @@ impl MonsterTemplate {
     fn velocity(mul: f32) -> Velocity {
         Velocity::new(5.0 * mul, Duration::from_secs(1))
     }
+
     fn damage(mul: f32) -> f32 {
         mul
     }
+
     fn reward(mul: usize) -> usize {
         mul
     }
+
     pub fn new(kind: MonsterKind) -> Self {
-        let (velocity, damage, reward, _skills): (f32, f32, usize, Vec<()>) = match kind {
-            MonsterKind::Mob01 => (1.0, 1.0, 3, vec![]),
-            MonsterKind::Mob02 => (1.0, 1.0, 3, vec![]),
-            MonsterKind::Mob03 => (1.0, 1.0, 3, vec![]),
-            MonsterKind::Mob04 => (1.0, 1.0, 3, vec![]),
-            MonsterKind::Mob05 => (1.0, 1.0, 3, vec![]),
-            MonsterKind::Mob06 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob07 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob08 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob09 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob10 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob11 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob12 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob13 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob14 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob15 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob16 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob17 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob18 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob19 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob20 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob21 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob22 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob23 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob24 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob25 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob26 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob27 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob28 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob29 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob30 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob31 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob32 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob33 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob34 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob35 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob36 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob37 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob38 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob39 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob40 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob41 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob42 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob43 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob44 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob45 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob46 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob47 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob48 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob49 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Mob50 => (1.0, 1.0, 5, vec![]),
-            MonsterKind::Boss01 => (1.0, 15.0, 50, vec![]),
-            MonsterKind::Boss02 => (1.0, 20.0, 75, vec![]),
-            MonsterKind::Boss03 => (1.0, 20.0, 100, vec![]),
-            MonsterKind::Boss04 => (1.0, 25.0, 100, vec![]),
-            MonsterKind::Boss05 => (1.0, 25.0, 125, vec![]),
-            MonsterKind::Boss06 => (1.0, 25.0, 125, vec![]),
-            MonsterKind::Boss07 => (1.0, 50.0, 125, vec![]),
-            MonsterKind::Boss08 => (1.0, 50.0, 125, vec![]),
-            MonsterKind::Boss09 => (1.0, 50.0, 125, vec![]),
-            MonsterKind::Boss10 => (1.0, 50.0, 125, vec![]),
-            MonsterKind::Boss11 => (1.0, 50.0, 125, vec![]),
-        };
+        Self::new_with_config(kind, &crate::config::GameConfig::default_config())
+    }
+
+    pub fn new_with_config(kind: MonsterKind, config: &crate::config::GameConfig) -> Self {
+        let stats = config
+            .monsters
+            .stats
+            .get(&kind)
+            .expect("missing monster stats for kind");
         Self {
             kind,
-            max_hp: Self::get_base_max_hp(kind),
+            max_hp: stats.base_hp,
             skills: vec![],
-            velocity: Self::velocity(velocity),
-            damage: Self::damage(damage),
-            reward: Self::reward(reward),
+            velocity: Self::velocity(stats.velocity_mul),
+            damage: Self::damage(stats.damage),
+            reward: Self::reward(stats.reward),
         }
     }
