+++ tower-defense/src/game_state/upgrade/generation/upgrade_candidate_table.rs
@@ -3,76 +3,94 @@ use crate::game_state::upgrade::MAX_REMOVE_NUMBER_RANKS;
 use crate::game_state::{GameState, tower::TowerKind};
 use rand::{Rng, seq::SliceRandom, thread_rng};
 
-type KindGen = fn() -> UpgradeKind;
+type KindGen = Box<dyn Fn() -> UpgradeKind + Send + Sync>;
 
 pub struct CandidateRow {
     pub weight: f32,
     pub kind_gen: KindGen,
 }
 
-pub fn generate_tower_damage_upgrade_candidate_table(_game_state: &GameState) -> Vec<CandidateRow> {
-    vec![
-        CandidateRow {
-            weight: 13.0,
-            kind_gen: || UpgradeKind::CainSword {
-                damage_multiplier: thread_rng().gen_range(1.15..1.5),
-            },
-        },
-        CandidateRow {
-            weight: 13.0,
-            kind_gen: || UpgradeKind::LongSword {
-                damage_multiplier: thread_rng().gen_range(1.15..1.5),
-            },
-        },
-        CandidateRow {
-            weight: 13.0,
-            kind_gen: || UpgradeKind::Mace {
-                damage_multiplier: thread_rng().gen_range(1.15..1.5),
-            },
-        },
-        CandidateRow {
-            weight: 13.0,
-            kind_gen: || UpgradeKind::ClubSword {
-                damage_multiplier: thread_rng().gen_range(1.15..1.5),
-            },
-        },
-        CandidateRow {
-            weight: 50.0,
-            kind_gen: || UpgradeKind::Spoon {
-                damage_multiplier: thread_rng().gen_range(1.3..1.75),
-            },
-        },
-        CandidateRow {
-            weight: 20.0,
-            kind_gen: || UpgradeKind::SingleChopstick {
-                damage_multiplier: thread_rng().gen_range(1.2..1.4),
-            },
-        },
-        CandidateRow {
-            weight: 20.0,
-            kind_gen: || UpgradeKind::PairChopsticks {
-                damage_multiplier: thread_rng().gen_range(1.2..1.4),
-            },
-        },
-        CandidateRow {
-            weight: 20.0,
-            kind_gen: || UpgradeKind::FountainPen {
-                damage_multiplier: thread_rng().gen_range(1.2..1.4),
-            },
-        },
-        CandidateRow {
-            weight: 20.0,
-            kind_gen: || UpgradeKind::Brush {
-                damage_multiplier: thread_rng().gen_range(1.2..1.4),
-            },
-        },
-        CandidateRow {
-            weight: 20.0,
-            kind_gen: || UpgradeKind::BrokenPottery {
-                damage_multiplier: thread_rng().gen_range(1.15..1.25),
-            },
-        },
-    ]
+pub fn generate_tower_damage_upgrade_candidate_table(game_state: &GameState) -> Vec<CandidateRow> {
+    game_state
+        .config
+        .upgrades
+        .tower_damage_upgrades
+        .iter()
+        .map(|(name, entry)| CandidateRow {
+            weight: entry.weight,
+            kind_gen: make_tower_damage_upgrade_kind_gen(name, entry.damage_multiplier_range),
+        })
+        .collect()
+}
+
+fn make_tower_damage_upgrade_kind_gen(
+    name: &str,
+    range: Option<(f32, f32)>,
+) -> KindGen {
+    let range = range.unwrap_or((1.0, 1.0));
+    match name {
+        "CainSword" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::CainSword {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "LongSword" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::LongSword {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "Mace" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::Mace {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "ClubSword" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::ClubSword {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "Spoon" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::Spoon {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "SingleChopstick" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::SingleChopstick {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "PairChopsticks" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::PairChopsticks {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "FountainPen" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::FountainPen {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "Brush" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::Brush {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "BrokenPottery" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::BrokenPottery {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        other => panic!("Unknown tower damage upgrade kind: {other}"),
+    }
 }
 
 pub fn generate_treasure_upgrade_candidate_table(game_state: &GameState) -> Vec<CandidateRow> {
@@ -91,68 +109,57 @@ pub fn generate_treasure_upgrade_candidate_table(game_state: &GameState) -> Vec<
         });
     };
 
-    push_row(
-        || UpgradeKind::Magnet,
-        Some((upgrade_state.gold_earn_plus, MAX_GOLD_EARN_PLUS)),
-        50.0,
-    );
-    push_row(
-        || UpgradeKind::Backpack,
-        Some((upgrade_state.shop_slot_expand, MAX_SHOP_SLOT_EXPAND)),
-        50.0,
-    );
-    push_row(
-        || UpgradeKind::DiceBundle,
-        Some((upgrade_state.dice_chance_plus, MAX_DICE_CHANCE_PLUS)),
-        10.0,
-    );
-    push_row(
-        || UpgradeKind::EnergyDrink,
-        Some((
-            upgrade_state.shop_item_price_minus,
-            MAX_SHOP_ITEM_PRICE_MINUS_UPGRADE,
-        )),
-        10.0,
-    );
-    push_row(
-        || UpgradeKind::PerfectPottery {
-            damage_multiplier: thread_rng().gen_range(1.3..1.75),
-        },
-        None,
-        25.0,
-    );
-    push_row(
-        || UpgradeKind::BrokenPottery {
-            damage_multiplier: thread_rng().gen_range(1.15..1.25),
-        },
-        None,
-        20.0,
-    );
-    push_row(
-        || UpgradeKind::FourLeafClover,
-        Some((upgrade_state.shorten_straight_flush_to_4_cards as usize, 1)),
-        10.0,
-    );
-    push_row(
-        || UpgradeKind::Rabbit,
-        Some((upgrade_state.skip_rank_for_straight as usize, 1)),
-        10.0,
-    );
-    push_row(
-        || UpgradeKind::BlackWhite,
-        Some((upgrade_state.treat_suits_as_same as usize, 1)),
-        10.0,
-    );
-    push_row(
-        || UpgradeKind::Eraser,
-        Some((
-            upgrade_state.removed_number_rank_count,
-            MAX_REMOVE_NUMBER_RANKS,
-        )),
-        10.0,
-    );
+    for (name, entry) in &game_state.config.upgrades.treasure_upgrades {
+        let weight = entry.weight;
+        let kind_gen = make_treasure_upgrade_kind_gen(name, entry.damage_multiplier_range);
+        let current_and_max = match name.as_str() {
+            "Magnet" => Some((upgrade_state.gold_earn_plus, MAX_GOLD_EARN_PLUS)),
+            "Backpack" => Some((upgrade_state.shop_slot_expand, MAX_SHOP_SLOT_EXPAND)),
+            "DiceBundle" => Some((upgrade_state.dice_chance_plus, MAX_DICE_CHANCE_PLUS)),
+            "EnergyDrink" => Some((
+                upgrade_state.shop_item_price_minus,
+                MAX_SHOP_ITEM_PRICE_MINUS_UPGRADE,
+            )),
+            "FourLeafClover" => Some((upgrade_state.shorten_straight_flush_to_4_cards as usize, 1)),
+            "Rabbit" => Some((upgrade_state.skip_rank_for_straight as usize, 1)),
+            "BlackWhite" => Some((upgrade_state.treat_suits_as_same as usize, 1)),
+            "Eraser" => Some((upgrade_state.removed_number_rank_count, MAX_REMOVE_NUMBER_RANKS)),
+            _ => None,
+        };
+        push_row(kind_gen, current_and_max, weight);
+    }
 
     rows
+}
+
+fn make_treasure_upgrade_kind_gen(
+    name: &str,
+    range: Option<(f32, f32)>,
+) -> KindGen {
+    let range = range.unwrap_or((1.0, 1.0));
+    match name {
+        "Magnet" => Box::new(|| UpgradeKind::Magnet),
+        "Backpack" => Box::new(|| UpgradeKind::Backpack),
+        "DiceBundle" => Box::new(|| UpgradeKind::DiceBundle),
+        "EnergyDrink" => Box::new(|| UpgradeKind::EnergyDrink),
+        "PerfectPottery" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::PerfectPottery {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "BrokenPottery" => {
+            let (min, max) = range;
+            Box::new(move || UpgradeKind::BrokenPottery {
+                damage_multiplier: thread_rng().gen_range(min..max),
+            })
+        }
+        "FourLeafClover" => Box::new(|| UpgradeKind::FourLeafClover),
+        "Rabbit" => Box::new(|| UpgradeKind::Rabbit),
+        "BlackWhite" => Box::new(|| UpgradeKind::BlackWhite),
+        "Eraser" => Box::new(|| UpgradeKind::Eraser),
+        other => panic!("Unknown treasure upgrade kind: {other}"),
+    }
 }
 
 #[allow(dead_code)]
